Ax 764 when transaction is rejected pop up should show according information (#517)
* chore: working on adding dialogs when the transaction is rejected

* chore: show transaction for when stake is rejected

* chore: reject transaction dialog shows for redeem dialog

* refactor: formatting

* refactor: separate failed dialog in its own file

* chore: formatting

* chore: closing all dialogs when the transaction is rejected

* refactor: removed unused imports
