+++ lib/pages/farm/components/stake_approve_button.dart
@@ -1,6 +1,6 @@
 import 'package:ax_dapp/pages/farm/modules/axl_info.dart';
 import 'package:ax_dapp/service/controller/farms/farm_controller.dart';
-import 'package:ax_dapp/service/dialog.dart';
+import 'package:ax_dapp/service/failed_dialog.dart';
 import 'package:ax_dapp/service/tracking/tracking_cubit.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -75,6 +75,14 @@ class _StakeApproveButtonState extends State<StakeApproveButton> {
             axlBalance: info.axlBalance,
           );
     }).catchError((_) {
+      showDialog<void>(
+        context: context,
+        builder: (context) => const FailedDialog(),
+      ).then((value) {
+        if (mounted) {
+          Navigator.pop(context);
+        }
+      });
       setState(() {
         isApproved = false;
         text = 'Approve';
@@ -111,8 +119,11 @@ class _StakeApproveButtonState extends State<StakeApproveButton> {
                 context: context,
                 builder: (BuildContext context) =>
                     widget.confirmDialog(context),
-              );
-
+              ).then((value) {
+                if (mounted) {
+                  Navigator.pop(context);
+                }
+              });
               final info = getStakeInfo();
               context.read<TrackingCubit>().onStakeSuccess(
                     tickerPair: info.tickerPair,
@@ -124,7 +135,11 @@ class _StakeApproveButtonState extends State<StakeApproveButton> {
               showDialog<void>(
                 context: context,
                 builder: (context) => const FailedDialog(),
-              );
+              ).then((value) {
+                if (mounted) {
+                  Navigator.pop(context);
+                }
+              });
             });
           } else {
             //Approve button was pressed
